pyproject.toml
[tool.poetry]
name = "utilibot"
version = "1.0.0"
description = "A discord bot in python which has many utilities"
authors = ["Tyman"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.*"
postbin = {git = "https://github.com/dragdev-studios/PostBin"}
python-dotenv = "^0.15.0"
jishaku = "^1.19.2"
youtube_dl = "^2020.11.19"
DiscordUtils = "^1.2.4"
aiofiles = "^0.6.0"
aiosqlite = "^0.16.0"
youtube-search = "^1.1.1"
dblpy = "^0.4.0"
pytz = "^2020.4"
dpytils = "^0.0.5"
psutil = "^5.7.3"
tabulate = "^0.8.7"
riftgun = "^1.1.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", extras = ["voice"]}
fastapi = "^0.62.0"
uvicorn = "^0.12.3"


[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
